code/Iris_end2end/requirements.txt
azure-cli==2.15.1
mlflow==1.17.0
databricks-cli==0.14.3
scikit-learn
azureml-sdk
kedro-viz
azureml-mlflow


